requirements.txt
Flask~=3.0
Flask-SQLAlchemy~=3.1
Flask-Migrate~=4.0
Flask-Login~=0.6
Flask-WTF~=1.2
flask-talisman~=1.1
Flask-Limiter~=3.5
gunicorn~=22.0
cryptography~=42.0
pyotp~=2.9
qrcode[pil]~=7.4
Pillow~=10.0
opencv-python-headless~=4.9
mediapipe~=0.10
scikit-image~=0.22
numpy~=1.26
duckduckgo-search~=5.0
reportlab~=4.1
pytest~=8.0
